Add realtime chat-messaging
